% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_sam.R
\name{calc_mase}
\alias{calc_mase}
\title{レトロの結果から各Indexに対する予測値を抽出して、Mean Absolute Scaled Errorを計算する関数}
\usage{
calc_mase(samres, retrores, h = 1, log = FALSE, index_name = NULL)
}
\arguments{
\item{samres}{SAM object}

\item{retrores}{\code{retro_sam(res,...)}で実地されたレトロ解析の結果オブジェクト}

\item{log}{MASEを計算するときに、Indexの観測値と予測値に対してlogを取るかどうか（デフォルトはFALSE）}

\item{index_name}{各Indexの名前ベクトル、Indexの数だけ必要}
}
\description{
レトロの結果から各Indexに対する予測値を抽出して、Mean Absolute Scaled Errorを計算する関数
}
