% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcu_BRs.R
\name{Calcu_Fmax}
\alias{Calcu_Fmax}
\title{Calculate the fishing mortality that maximizes the yield per recruit.}
\usage{
Calcu_Fmax(M, Sel, w, g, A = 6, F_max = 10, method = "Baranov")
}
\arguments{
\item{M}{A vector of natural mortality rate at age a. The length should be A+1.}

\item{Sel}{A vector of selectivity at age a. The length should be A+1.}

\item{w}{A vector of weight at age a. The length should be A+1.}

\item{g}{A vector of maturity at age a. The length should be A+1.}

\item{A}{Plus group age. Default is 6.}

\item{F_max}{Maximum value of F values to be searched for determining F\%SPR.
This value should be large enough so that the true F\%SPR value be included. Default is 10}

\item{method}{Method of the fishing equation. Either "Baranov" or "Pope" is allowed. Default is "Baranov".}
}
\value{
A value of the F that maximizes the yield per recruit.
}
\description{
\emph{Calcu_Fmax} calculates the fishing mortality that maximizes the yield per recruit.
}
