% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcu_BRs.R
\name{Calcu_Fmsy}
\alias{Calcu_Fmsy}
\title{Calculate the fishing mortality that maximizes the maximum sustainable yield (Fmsy) with fixed steepness parameters (h) in the Bevertoh-Holt stock-recruitment relationship.}
\usage{
Calcu_Fmsy(M, Sel, w, g, A = 6, method, F_max = 10, alpha, beta, method_SR)
}
\arguments{
\item{M}{A vector of natural mortality rate at age a. The length should be A+1.}

\item{Sel}{A vector of selectivity at age a. The length should be A+1.}

\item{w}{A vector of weight at age a. The length should be A+1.}

\item{g}{A vector of maturity at age a. The length should be A+1.}

\item{A}{Plus group age. Default is 6.}

\item{method}{Method of the fishing equation. Either "Baranov" or "Pope" is allowed.}

\item{F_max}{Maximum value of F values to be searched for determining Fmsy.}

\item{alpha}{Estimated Parameter alpha in the Beverton-Holt stock recruitment relationship.}

\item{beta}{Estimated Parameter beta in the Beverton-Holt stock recruitment relationship.}
}
\value{
A value of the F that maximizes the yield per recruit.
}
\description{
\emph{Calcu_Fmsy} calculates the fishing mortality that maximizes the maximum sustainable yield (Fmsy) with fixed steepness parameters (h).
}
