% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcu_BRs.R
\name{Calcu_N}
\alias{Calcu_N}
\title{Calculate the relative number at age at the equilibrium.}
\usage{
Calcu_N(Fish_mort, M, Sel, A = 6)
}
\arguments{
\item{Fish_mort}{Fishing mortality to calculate SBR. Should be provided with a numeric value.}

\item{M}{A vector of natural mortality rate at age a. The length should be A+1.}

\item{Sel}{A vector of selectivity at age a. The length should be A+1.}

\item{A}{Plus group age. Default is 6.}
}
\description{
\emph{Calcu_N} calculates the relative number at age at the equilibrium (N) with a given fishing mortality value.
}
