% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcu_BRs.R
\name{Calcu_SBR}
\alias{Calcu_SBR}
\title{Calculate spawning biomass per recruit (SBR) with a given fishing mortality value.}
\usage{
Calcu_SBR(Fish_mort, M, Sel, w, g, A = 6)
}
\arguments{
\item{Fish_mort}{Fishing mortality to calculate SBR. Should be provided with a numeric value.}

\item{M}{A vector of natural mortality rate at age a. The length should be A+1.}

\item{Sel}{A vector of selectivity at age a. The length should be A+1.}

\item{w}{A vector of weight at age a. The length should be A+1.}

\item{g}{A vector of maturity at age a. The length should be A+1.}

\item{A}{Plus group age. Default is 6.}
}
\value{
A value of spawing biomass par recruit with the given fishing mortality.
}
\description{
\emph{Calcu_SBR} calculates SBR from the statistics derived from the \emph{Calcu_stat} function and input data with a given fishing mortality value.
}
