% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcu_BRs.R
\name{Calcu_SPR_X}
\alias{Calcu_SPR_X}
\title{Calculate F\%SPR: the fishing mortality that reduces the spawning biomass per recruit (SBR) to X \% of those without fishing (SPR0).}
\usage{
Calcu_SPR_X(x, M, Sel, w, g, A = 6, F_max = 2)
}
\arguments{
\item{x}{The proportion (X) to calculate F\% SPR. This parameter should be provided as a proportion (i.e. within the range from 0 to 1). Defalut is 0.2.}

\item{M}{A vector of natural mortality rate at age a. The length should be A+1.}

\item{Sel}{A vector of selectivity at age a. The length should be A+1.}

\item{w}{A vector of weight at age a. The length should be A+1.}

\item{g}{A vector of maturity at age a. The length should be A+1.}

\item{A}{Plus group age. Default is 6.}

\item{F_max}{This value should be large enough so that the true F\%SPR value be included (otherwise a warning message appears). Defalut is 1.}
}
\value{
A value of F\% SPR.
}
\description{
\emph{Calcu_SPR_X} calculates F\%SPR (the fishing mortality that reduce the spawning biomass per recruit (SBR) to X \% of those without fishing (SPR0)).
}
