% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcu_BRs.R
\name{Calcu_YPR}
\alias{Calcu_YPR}
\title{Calculate yield per recruit with a given fishing mortality value.}
\usage{
Calcu_YPR(Fish_mort, M, Sel, w, g, A = 6, method = "Baranov")
}
\arguments{
\item{Fish_mort}{Fishing mortality to calculate yield per recruit. Should be provided with a numeric value.}

\item{M}{A vector of natural mortality rate at age a. The length should be A+1.}

\item{Sel}{A vector of selectivity at age a. The length should be A+1.}

\item{w}{A vector of weight at age a. The length should be A+1.}

\item{g}{A vector of maturity at age a. The length should be A+1.}

\item{A}{Plus group age. Default is 6.}

\item{method}{Method of the fishing equation. Either "Baranov" or "Pope" is allowed. Default is "Baranov".}
}
\value{
A value of yield per recruit with the given fishing mortality.
}
\description{
\emph{Calcu_YPR} calculates yield per recruit from the statistics derived from the \emph{Calcu_stat} function and input data with a given fishing mortality value.
}
