% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{divide_sigma}
\alias{divide_sigma}
\title{観測誤差とプロセス誤差をどこかの年齢間で分けて推定し直す関数}
\usage{
divide_sigma(
  samres,
  var = c("varC", "varF", "varN", "index.key")[1],
  which = (1:6)[1]
)
}
\arguments{
\item{samres}{sam object}

\item{var}{分ける誤差の指定. "varC"（年齢別漁獲尾数の観測誤差）,"varF"（Fのプロセス誤差）,"varN"（Nのプロセス誤差）,"index.key"（指標の観測誤差）のいずれかを選択．}

\item{which}{どこで区切りをいれるか. 1だと0歳と1歳以上で分け（加入が0歳の場合）、指標値の場合は1本目と2本目以降の観測誤差を分ける}
}
\description{
観測誤差とプロセス誤差をどこかの年齢間で分けて推定し直す関数
}
