% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_jitter.R
\name{do_jitter}
\alias{do_jitter}
\title{Do jitter analysis}
\usage{
do_jitter(samres, SD = 0.1, nsim = 100, seed = 1)
}
\arguments{
\item{samres}{sam object}

\item{SD}{SD for random normal distribution to be added to initial values}

\item{nsim}{default: 100}
}
\description{
Do jitter analysis
}
