% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_osa_Resid.R
\encoding{UTF-8}
\name{do_osa_resid}
\alias{do_osa_resid}
\title{SAMのOSA residualを計算する関数}
\usage{
do_osa_resid(
  samres,
  method = "oneStepGaussianOffMode",
  subset = NULL,
  trace = 2
)
}
\arguments{
\item{samres}{sam object}

\item{method}{Method to calculate OSA (see details).}

\item{subset}{Index vector of observations that will be added one by one during OSA. By default \code{1:length(observations)} (with \code{conditional} subtracted).}

\item{trace}{Logical; Trace progress? More options available for \code{method="oneStepGeneric"} - see details.}
}
\description{
SAMのOSA residualを計算する関数
}
