% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_factors.R
\name{.add_decomposition_summaries}
\alias{.add_decomposition_summaries}
\title{Add age-aggregated and percentage summaries to a decomposition}
\usage{
.add_decomposition_summaries(x)
}
\description{
Add age-aggregated and percentage summaries to a decomposition
}
\keyword{internal}
