% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_factors.R
\name{.as_age_year_matrix}
\alias{.as_age_year_matrix}
\title{Convert vector, matrix, or data.frame to age-year matrix}
\usage{
.as_age_year_matrix(x, target, name = "M")
}
\description{
Convert vector, matrix, or data.frame to age-year matrix
}
\keyword{internal}
