% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{.as_numeric_matrix}
\alias{.as_numeric_matrix}
\title{Convert matrix or data.frame to numeric matrix}
\usage{
.as_numeric_matrix(x, name = deparse(substitute(x)))
}
\description{
Convert matrix or data.frame to numeric matrix
}
\keyword{internal}
