% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_factors.R
\name{.permutations}
\alias{.permutations}
\title{Generate all permutations of a character vector}
\usage{
.permutations(x)
}
\description{
Generate all permutations of a character vector
}
\keyword{internal}
