% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_factors.R
\name{.shapley_transition}
\alias{.shapley_transition}
\title{Shapley decomposition for one transition}
\usage{
.shapley_transition(
  N_from,
  w_from,
  w_to,
  F_from,
  M_from,
  process_multiplier,
  m_from = NULL,
  m_to = NULL
)
}
\description{
Shapley decomposition for one transition
}
\keyword{internal}
