% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_factors.R
\name{.weighted_mortality}
\alias{.weighted_mortality}
\title{Average mortality through survival, weighted by numbers at age}
\usage{
.weighted_mortality(x, w)
}
\description{
Average mortality through survival, weighted by numbers at age
}
\keyword{internal}
