% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samprofile.r
\encoding{UTF-8}
\name{est_fixed}
\alias{est_fixed}
\title{SAMの固定効果だけを推定する関数}
\usage{
est_fixed(
  tmbdata,
  par_init,
  map,
  random = "U",
  cpp.file.name = "sam2",
  silent = TRUE,
  ...
)
}
\arguments{
\item{tmbdata}{tmbdata}

\item{par_init}{パラメータの初期値}

\item{map}{推定しない（初期値で固定する）パラメータ}

\item{random}{ランダム効果で推定するパラメータ}

\item{cpp.file.name}{cppファイルの名前（初期設定："sam"）}
}
\description{
SAMの固定効果だけを推定する関数
}
