% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samprofile.r
\encoding{UTF-8}
\name{est_mixed}
\alias{est_mixed}
\title{TMB::MakeADFunに必要な引数から固定効果とランダム効果を推定する関数}
\usage{
est_mixed(
  tmbdata,
  par_init,
  map,
  random = "U",
  cpp.file.name = "sam2",
  silent = TRUE,
  bias_correct = TRUE
)
}
\arguments{
\item{tmbdata}{tmbdata}

\item{par_init}{パラメータの初期値}

\item{map}{推定しない（初期値で固定する）パラメータ}

\item{random}{ランダム効果で推定するパラメータ}

\item{cpp.file.name}{cppファイルの名前（初期設定："sam"）}

\item{bias_correct}{平均値のバイアス補正をするかどうか}
}
\description{
TMB::MakeADFunに必要な引数から固定効果とランダム効果を推定する関数
}
