% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim_vpasam.R
\name{fit2PSdata}
\alias{fit2PSdata}
\title{生成された疑似データにVPA/SAMを推定し、Self-test/Cross-testを実行}
\usage{
fit2PSdata(res, PSdata)
}
\arguments{
\item{res}{フィットさせるSAM/VPAのオブジェクト}

\item{PSdata}{\code{popsim_vpa()}で生成された疑似データ}
}
\description{
生成された疑似データにVPA/SAMを推定し、Self-test/Cross-testを実行
}
