% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{index_plot}
\alias{index_plot}
\title{Indexの当てはまりについてプロットする関数}
\usage{
index_plot(
  samvpa_list,
  model_name = NULL,
  fleet_no = NULL,
  scales = c("free", "free_x", "free"),
  base_size = 16
)
}
\description{
Indexの当てはまりについてプロットする関数
}
