% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{index_plot2}
\alias{index_plot2}
\title{Indexの当てはまりについてプロットする関数}
\usage{
index_plot2(
  samres,
  index_name = NULL,
  nrow = 2,
  scales = c("free", "free_x", "free"),
  base_size = 14
)
}
\description{
Indexの当てはまりについてプロットする関数
}
