% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sakaios.r
\name{make_assess_result}
\alias{make_assess_result}
\title{SAMの結果を読み込んで主要なパラメータの予測値とSD、信頼区間などを出力する}
\usage{
make_assess_result(res = res, CI = 0.9)
}
\description{
SAMの結果を読み込んで主要なパラメータの予測値とSD、信頼区間などを出力する
}
