% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{plot_boosam}
\alias{plot_boosam}
\title{ブートストラップについてプロットする関数}
\usage{
plot_boosam(
  samres,
  boores,
  CI = 0.95,
  what.plot = c("biomass", "SSB", "Recruitment", "U"),
  draw_deltaCI = FALSE,
  scenario_name = c("Estiamte", "Bootstrap"),
  alpha = 0.4,
  size = 1,
  base_size = 14,
  log_scale = FALSE,
  legend_name = "Scenario",
  legend_nrow = 1,
  legend_position = "top",
  years = NULL,
  ncol = 2
)
}
\arguments{
\item{samres}{\code{sam()}の結果オブジェクト}

\item{boores}{\code{boo_sam}の結果オブジェクト}

\item{CI}{Confidence interval width. Set 0 to omit confidence intervals.}

\item{what.plot}{Statistics to plot.}

\item{draw_deltaCI}{デルタ法の信頼区間を描くか（デフォルト:FALSE）}

\item{scenario_name}{推定値の結果、ブートストラップの結果の凡例に使う名前}

\item{alpha}{Alpha value for confidence interval ribbons.}

\item{size}{Line width.}

\item{base_size}{Base font size.}

\item{log_scale}{If \code{TRUE}, use a log scale on the y-axis.}

\item{legend_name}{Legend title.}

\item{legend_nrow}{Number of rows in the legend.}

\item{legend_position}{Legend position.}

\item{years}{Years to include in the plot.}

\item{ncol}{Number of columns in the facet plot.}
}
\description{
ブートストラップについてプロットする関数
}
