% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{plot_cond_nll}
\alias{plot_cond_nll}
\title{条件付き負の対数尤度を成分別に描画する}
\usage{
plot_cond_nll(x, show_value = TRUE)
}
\arguments{
\item{x}{[get_cond_nll()] が返すデータフレーム。}

\item{show_value}{棒の外側に負の対数尤度を小数第1位まで表示するか。
既定値は `TRUE`。}
}
\value{
`ggplot` オブジェクト。
}
\description{
[get_cond_nll()] の出力を横向きの棒グラフで表示します。
}
\examples{
\dontrun{
data("samres_example", package = "frasam")
plot_cond_nll(get_cond_nll(samres))
}

}
