% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\encoding{UTF-8}
\name{plot_hindcastCV}
\alias{plot_hindcastCV}
\title{レトロの結果を使って資源量指標値に対するhindcast cross validationをプロットする関数}
\usage{
plot_hindcastCV(
  samres,
  retrores,
  h = 1,
  log = FALSE,
  index_name = NULL,
  show_mase = TRUE,
  mase_position = "upperright",
  use_index = NULL,
  years = NULL
)
}
\arguments{
\item{samres}{SAM object}

\item{retrores}{\code{retro_sam(res,...)}で実地されたレトロ解析の結果オブジェクト}

\item{log}{MASEを計算するときに、Indexの観測値と予測値に対してlogを取るかどうか（デフォルトはFALSE）}

\item{index_name}{各Indexの名前ベクトル、Indexの数だけ必要}

\item{show_mase}{MASEの結果を載せるかどうか}

\item{use_index}{特定のIndexを使う場合、\code{use_index = c(1,3)}のように指定する}

\item{years}{プロットする期間を指定する場合、\code{years = 2015:2024}のように指定する}
}
\description{
レトロの結果を使って資源量指標値に対するhindcast cross validationをプロットする関数
}
