% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\encoding{UTF-8}
\name{plot_osa_resid}
\alias{plot_osa_resid}
\title{OSA residualをプロットする関数}
\usage{
plot_osa_resid(osares)
}
\arguments{
\item{osares}{\code{do_osa_resid}の結果オブジェクト}
}
\description{
OSA residualをプロットする関数
}
