% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\encoding{UTF-8}
\name{plot_popsim}
\alias{plot_popsim}
\title{Popsimの結果ををプロットする関数}
\usage{
plot_popsim(
  res_true,
  fit2PS,
  CI = 0.95,
  what.plot = c("biomass", "SSB", "Recruitment", "U"),
  Age = NULL,
  scenario_name = c("True", "Simulation"),
  alpha = 0.4,
  size = 1,
  base_size = 14,
  log_scale = FALSE,
  legend_name = "Scenario",
  legend_nrow = 1,
  legend_position = "top",
  years = NULL,
  ncol = 2,
  sim_colour = "blue"
)
}
\arguments{
\item{res_true}{真の推定値をもつSAMまたはVPAのオブジェクト}

\item{fit2PS}{\code{fit2PSdata()} で得られる、疑似データにフィットさせた結果オブジェクト}
}
\description{
Popsimの結果ををプロットする関数
}
