% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{plot_SR_simple}
\alias{plot_SR_simple}
\title{再生産関係についてプロットする関数}
\usage{
plot_SR_simple(samres, length = 1000, base_size = 14, ...)
}
\arguments{
\item{samres}{SAMの結果オブジェクト}
}
\description{
再生産関係についてプロットする関数
}
