% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{retro_plot}
\alias{retro_plot}
\title{レトロスペクティブ解析の結果をプロットする}
\usage{
retro_plot(
  res,
  retro_res,
  start_year = NULL,
  scale = 1000,
  forecast = FALSE,
  base_size = 14,
  plot_mohn = TRUE,
  mohn_position = "upperleft"
)
}
\arguments{
\item{res}{sam object}

\item{retro_res}{\code{retro_sam}の結果}

\item{start_year}{プロットを開始する年}
}
\description{
レトロスペクティブ解析の結果をプロットする
}
