% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_prec.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{Simulate multivariate normal variables given a mean vector and precision matrix}
\usage{
rmvnorm_prec(mu, prec, n.sims, seed = 123)
}
\arguments{
\item{mu}{vector of parameter means}

\item{prec}{joint precision matrix}

\item{n.sims}{number of draws}

\item{seed}{seed number}
}
\value{
length(mu) by n.sims matrix of parameter draws
}
\description{
Simulate multivariate normal variables given a mean vector and precision matrix
}
