% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_do_call.R
\encoding{UTF-8}
\name{safe_do_call}
\alias{safe_do_call}
\title{do.callの引数が与えられていない場合にディフォルト値を与えて計算する関数}
\usage{
safe_do_call(func, args_list, default_args = list())
}
\arguments{
\item{func}{関数名}

\item{args_list}{引数名}
}
\description{
余計な引数があった場合、その関数を削除して実行する
}
\examples{
example_func <- function(a, b = 2, c = 3) {
  a + b + c
}

args <- list(a = 1, d = 4)  # `d` is an extra argument
safe_do_call(example_func, args)  # Returns 6 (1 + 2 + 3)

}
