% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samprofile.r
\encoding{UTF-8}
\name{samprofile}
\alias{samprofile}
\title{SAMで推定されたパラメータに対してプロファイル尤度を計算する}
\usage{
samprofile(
  samres,
  param_name,
  which_param = 1,
  param_range = NULL,
  length = 50
)
}
\arguments{
\item{samres}{sam object}

\item{param_name}{変化させるパラメータの名前}

\item{which_param}{変化させるパラメータの位置（複数のパラメータが同じ名前を持つときに使用）}
}
\description{
SAMで推定されたパラメータに対してプロファイル尤度を計算する
}
