% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim_vpasam.R
\encoding{UTF-8}
\name{sumup_popsim}
\alias{sumup_popsim}
\title{Self-test, Cross-testの結果をまとめるための関数}
\usage{
sumup_popsim(res_true, fit2PS, percent = FALSE, CI = 0.95, ...)
}
\arguments{
\item{res_true}{真の推定値をもつSAMまたはVPAのオブジェクト}

\item{fit2PS}{\code{fit2PSdata()} で得られる、疑似データにフィットさせた結果オブジェクト}

\item{percent}{Logical; if \code{TRUE} (default), express relative metrics (RMSRE, MARE, MedRelBias)
in percentage (i.e., multiplied by 100). If \code{FALSE}, leave as proportions.}
}
\value{
A list with the following components:
\itemize{
  \item \code{summary}: A data.frame containing performance metrics.
    See [calc_metrics()] for details on the definitions of RMSE, MAE, R2_model, etc.
  \item \code{all}: すべてのRunについての結果のデータフレーム
}
}
\description{
Self-test, Cross-testの結果をまとめるための関数
}
