% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sam.R
\name{update_sam}
\alias{update_sam}
\title{Update SAM result by new parameters of fixed and random effects}
\usage{
update_sam(samres, new_par)
}
\arguments{
\item{samres}{SAM object}

\item{new_par}{new parameter set}
}
\description{
Update SAM result by new parameters of fixed and random effects
}
