% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sam_tmb.R
\encoding{UTF-8}
\name{use_sam_tmb}
\alias{use_sam_tmb}
\title{SAMでTMBで実行するためにcppファイルのコンパイル等をする関数}
\usage{
use_sam_tmb(
  TmbFile = "sam2",
  CppDir = system.file("executable", package = "frasam"),
  RunDir = getwd(),
  overwrite = FALSE,
  compile = c("auto", "always", "never"),
  auto_update = NULL
)
}
\arguments{
\item{TmbFile}{Cppファイルの名前}

\item{CppDir}{Cppファイルが格納されているディレクトリ}

\item{RunDir}{CppファイルとDLLを配置するディレクトリ}

\item{overwrite}{RunDirのCppファイルを上書きするかどうか}

\item{compile}{DLLをコンパイルするかどうか。"auto"ではCppファイルがDLLより新しい場合にコンパイルする}

\item{auto_update}{互換性のために残している引数。compile = "auto"を使用してください}
}
\description{
SAMでTMBで実行するためにcppファイルのコンパイル等をする関数
}
\examples{
\dontrun{
use_sam_tmb()
}

}
