% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sam_tibble.R
\encoding{UTF-8}
\name{convert_sam_tibble}
\alias{convert_sam_tibble}
\title{SAMの結果オブジェクトをtibble形式に変換する関数}
\usage{
convert_sam_tibble(samres)
}
\arguments{
\item{vpares}{vpaの結果のオブジェクト}
}
\description{
SAMの結果オブジェクトをtibble形式に変換する関数
}
