% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\encoding{UTF-8}
\name{get_pm}
\alias{get_pm}
\title{utilities for extracting statistics}
\usage{
get_pm(
  res_sam,
  res_future = NULL,
  waa_catch = res_sam$input$dat$waa,
  last_year = 2022,
  waa_biom = res_sam$input$dat$waa,
  year_biol = 2020:2022,
  year_Fcur = 2020:2022,
  perSPR = c(30, 40, 50)
)
}
\arguments{
\item{res_sam}{result of sam function}

\item{res_future}{result of future projection}

\item{waa_catch}{weight at age for calculating total catch weight. This matrix should be given because SAM doesn't use this biological parameters}

\item{last_year}{last year of stock assessment}

\item{waa_biom}{weight at age for calculating biomass weight. If this is not given, sam_res$input$dat$waa is automatically used}

\item{year_biol}{year range to derive biological parameters. Biological parameters are averaged by age during the given period}

\item{year_Fcur}{year range to derive F at age considred as F current. Fs at age are averaged by age during the given period}

\item{perSPR}{percent SPR (%) for calculating F%SPR}
}
\description{
utilities for extracting statistics
}
\examples{
\donotrun{
res_pm_all <- purrr::map_dfr(basecase_list,
                  function(x) get_pm(x, NULL, waa_catch=x$input$dat$waa), .id="id") \%>\%
                  pivot_wider(values_from=value, names_from=stat)
}

}
