% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\encoding{UTF-8}
\name{plot_samvpa}
\alias{plot_samvpa}
\title{SAM or VPAの結果を描くグラフ}
\usage{
plot_samvpa(
  vpa_sam_list,
  CI = 0.95,
  scenario_name = NULL,
  alpha = 0.4,
  size = 1,
  base_size = 16,
  log_scale = FALSE,
  legend_name = "Scenario",
  legend_nrow = 1,
  legend_position = "top",
  what.plot = c("biomass", "SSB", "Recruitment", "U")
)
}
\arguments{
\item{samres}{samの結果オブジェクト}
}
\description{
SAM or VPAの結果を描くグラフ
}
