% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_sam.R
\name{retro_plot}
\alias{retro_plot}
\title{レトロスペクティブ解析の結果をプロットする}
\usage{
retro_plot(
  res,
  retro_res,
  start_year = NULL,
  scale = 1000,
  forecast = FALSE,
  base_size = 16,
  plot_mohn = TRUE,
  mohn_position = "upperleft"
)
}
\arguments{
\item{res}{フルデータを使った解析結果 (vpa or sam)}

\item{retro_res}{}
}
\description{
レトロスペクティブ解析の結果をプロットする
}
