% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_sam.R
\name{retro_sam}
\alias{retro_sam}
\title{レトロスペクティブ解析を実施する}
\usage{
retro_sam(
  res,
  n = 5,
  stat = "mean",
  b.fix = TRUE,
  remove_short_index = TRUE,
  map_add = NULL,
  p0_retro_list = NULL
)
}
\description{
レトロスペクティブ解析を実施する
}
