% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_sam.R
\name{retro_sam}
\alias{retro_sam}
\title{レトロスペクティブ解析を実施する}
\usage{
retro_sam(
  res,
  n = 5,
  stat = "mean",
  b.fix = TRUE,
  remove_short_index = -1,
  map_add = NULL,
  p0_retro_list = NULL
)
}
\arguments{
\item{res}{SAM object}

\item{n}{the number of peels}
}
\description{
retrospective forecasting is also possible
}
