% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{SAMによる資源計算を実施する}
\usage{
sam(
  dat,
  last.catch.zero = FALSE,
  cpp.file.name = "sam",
  tmb.run = FALSE,
  abund = c("B"),
  rec.age = 0,
  min.age = rep(0, length(abund)),
  max.age = rep(0, length(abund)),
  plus.group = TRUE,
  alpha = 1,
  b.est = FALSE,
  b.fix = rep(NA, length(abund)),
  varC = 0,
  varN = 0,
  varF = 0,
  varN.fix = NULL,
  est.method = "ml",
  SR = "BH",
  AR = 0,
  rho.mode = 2,
  q.init = NULL,
  sdFsta.init = NULL,
  sdLogN.init = NULL,
  sdLogObs.init = NULL,
  rho.init = NULL,
  a.init = NULL,
  b.init = NULL,
  ref.year = 1:5,
  bias.correct = TRUE,
  bias.correct.sd = FALSE,
  get.random.vcov = FALSE,
  silent = FALSE,
  remove.Fprocess.year = NULL,
  RW.Forder = 0,
  map.add = NULL,
  p0.list = NULL,
  scale = 1000,
  gamma = 10,
  sel.def = "max",
  use.index = NULL,
  upper = NULL,
  lower = NULL
)
}
\arguments{
\item{dat}{samに使用するdataでrvpaと同じフォーマットで利用可能}

\item{rec.age}{加入年齢 (default: 0)}

\item{min.age}{Indexの最低年齢 (\code{frasyr::vpa()}と同じで最小の年齢を0とする)}

\item{max.age}{Indexの最高年齢 (\code{frasyr::vpa()}と同じで最小の年齢を0とする)}

\item{SR}{再生産関係："RW", "BH", "RI", "HS", "Mesnil", or "Const"}
}
\description{
SAMによる資源計算を実施する
}
