% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_sigma_grid}
\alias{select_sigma_grid}
\title{観測誤差やプロセス誤差のステップ形式のモデル選択（複数の変数について）}
\usage{
select_sigma_grid(
  samres,
  grid = expand.grid(var = c("varC", "varF", "varN", "index.key")[1:2], X = 1:2),
  stopAIC = TRUE
)
}
\arguments{
\item{samres}{sam object}

\item{grid}{'var'と'X'からなるdata.frame}

\item{stopAIC}{AICが小さくならなかった時点で計算をやめるか（default: TRUE)}
}
\description{
\code{devide_sigma}関数を順々に実行し、AIC規準で最適なモデルを探索
}
