% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sam_tmb.R
\encoding{UTF-8}
\name{use_sam_tmb}
\alias{use_sam_tmb}
\title{SAMでTMBで実行するためにcppファイルのコンパイル等をする関数}
\usage{
use_sam_tmb(
  TmbFile = "sam2",
  CppDir = system.file("executable", package = "frasam"),
  RunDir = getwd(),
  overwrite = FALSE,
  compile = c("auto", "always", "never"),
  auto_update = NULL,
  ...
)
}
\arguments{
\item{TmbFile}{Cppファイルの名前}

\item{CppDir}{Cppファイルが格納されているディレクトリ}
}
\description{
SAMでTMBで実行するためにcppファイルのコンパイル等をする関数
}
\examples{
\dontrun{
use_sam_tmb()
}

}
